updated implementation and integration tests
