+++ src/main/java/com/laserfiche/repository/api/clients/impl/RepositoriesClientImpl.java
@@ -1,12 +1,15 @@
 package com.laserfiche.repository.api.clients.impl;
 
 import com.laserfiche.api.client.deserialization.ProblemDetailsDeserializer;
+import com.laserfiche.api.client.deserialization.TokenClientObjectMapper;
 import com.laserfiche.api.client.httphandlers.HttpRequestHandler;
 import com.laserfiche.api.client.model.ApiException;
 import com.laserfiche.api.client.model.ProblemDetails;
 import com.laserfiche.repository.api.clients.RepositoriesClient;
 import com.laserfiche.repository.api.clients.impl.model.RepositoryInfo;
 import kong.unirest.HttpResponse;
+import kong.unirest.ObjectMapper;
+import kong.unirest.Unirest;
 import kong.unirest.UnirestInstance;
 import kong.unirest.json.JSONArray;
 import kong.unirest.json.JSONObject;
@@ -50,5 +53,50 @@ public class RepositoriesClientImpl extends ApiClient implements RepositoriesCli
         };
         return ApiClientUtils.sendRequestWithRetry(httpClient, httpRequestHandler, baseUrl + "/v1/Repositories", "GET",
                 null, null, null, null, null, null, new HashMap<String, String>(), false, parseResponse);
+    }
+
+    /**
+     * Returns the repository resource list that current user has access to given the API server base URL. Only available in Laserfiche Self-Hosted.
+     *
+     * @param url API server base URL e.g., https://{APIServerName}/LFRepositoryAPI
+     * @return Get the respository resource list successfully.
+     */
+    public static RepositoryInfo[] getSelfHostedRepositoryList(String url) {
+        Map<String, String> headerKeyValuePairs = new HashMap<>();
+        headerKeyValuePairs.put("accept", "application/json");
+        HttpResponse<Object> httpResponse = null;
+        String responseJson;
+        if (url.endsWith("/")) {
+            url = url.substring(0, url.length() - 1);
+        }
+        String baseUrl = url + "/v1/Repositories";
+        ObjectMapper objectMapper = new TokenClientObjectMapper();
+        UnirestInstance httpClient = Unirest.spawnInstance();
+        httpClient
+                .config()
+                .setObjectMapper(objectMapper);
+        httpResponse = httpClient
+                .get(baseUrl)
+                .headers(headerKeyValuePairs)
+                .asObject(Object.class);
+        Object body = httpResponse.getBody();
+        Map<String, String> headersMap = ApiClientUtils.getHeadersMap(httpResponse.getHeaders());
+        if (httpResponse.getStatus() == 200) {
+            try {
+                responseJson = new JSONArray(((ArrayList) body).toArray()).toString();
+                return objectMapper.readValue(responseJson, RepositoryInfo[].class);
+            } catch (Exception e) {
+                throw ApiException.create(httpResponse.getStatus(), headersMap, null, e);
+            }
+        } else {
+            ProblemDetails problemDetails;
+            try {
+                String jsonString = new JSONObject(body).toString();
+                problemDetails = ProblemDetailsDeserializer.deserialize(objectMapper, jsonString);
+            } catch (Exception e) {
+                throw ApiException.create(httpResponse.getStatus(), headersMap, null, e);
+            }
+            throw ApiClientUtils.createApiException(httpResponse, problemDetails);
+        }
     }
 }
